fix: Fixed naming conflicts in files
